+++ CHANGELOG.md
@@ -12,11 +12,20 @@
 - Thematic mode added - it allows you to play with words of specific categories.
 - Custom attempts added - the user can now set the number of guesses they want.
 ### Visual changes:
-- Board and options added.
+- Menu and options added.
 - Improved display of HOW TO PLAY and win/lose messages.
 ### Internal Changes:
 - Animals, Countries and fr-vg lists added for thematic mode.
 - Valid wordle list modified and slightly extended.
 - Fixed minor bugs.
 
-
+## [1.2.0]
+### New Features:
+- Green mode only - The feedback is limited, only green letters will be displayed.
+- Strict mode - You won't be able to use letters you know are not in the word, so as you go on playing, it becomes more difficult.
+### Visual changes:
+- Your attempts are now displayed.
+- The modes are toggleable.
+### Internal Changes:
+- Color list added for thematic mode.
+- Valid wordle list modified and slightly extended.